+++ esp_cam_sensor/CHANGELOG.md
@@ -7,6 +7,7 @@
   - YUV422 supports two byte orders: YUYV and UYVY, represented by `ESP_CAM_SENSOR_PIXFORMAT_YUV422_YUYV` and `ESP_CAM_SENSOR_PIXFORMAT_YUV422_UYVY` respectively; RGB565 supports two byte orders: RGB565_LE and RGB565_BE, represented by `ESP_CAM_SENSOR_PIXFORMAT_RGB565_LE` and `ESP_CAM_SENSOR_PIXFORMAT_RGB565_BE` respectively.
 - Added the 2-bit data I/O mode for the parallel I/O interface of the SPI camera interface driver.
   - The parallel I/O interface is supported in ESP-IDF version 5.5.2 and later.
+- Added OV9281 MIPI camera sensor driver.
 
 ## 1.7.0